Fix integer overflow in monitoring counter accumulation
The getCount() and getAllCounts() functions accumulated counter values
into an unsigned int, then cast to int for the return value. This could
cause incorrect results when the sum exceeds INT_MAX (2,147,483,647):
- Undefined behavior from signed integer overflow
- Negative values returned for large counter sums

Fix by using long long for accumulation and removing the int cast.
AmArg already supports long long int values, so this change is
API-compatible while correctly handling counter sums up to 9.2e18.

+++ apps/monitoring/Monitoring.cpp
@@ -334,7 +334,7 @@ void Monitor::getCount(const AmArg& args, AmArg& ret) {
     gettimeofday(&to, NULL);
   }
 
-  unsigned int res = 0;
+  long long res = 0;
 
   LogBucket& bucket = getLogBucket(args[0].asCStr());
   bucket.log_lock.lock();
@@ -369,7 +369,7 @@ void Monitor::getCount(const AmArg& args, AmArg& ret) {
   }
   bucket.log_lock.unlock();
 
-  ret.push((int)res);
+  ret.push(res);
 }
 
 // Expected args:
@@ -426,7 +426,7 @@ void Monitor::getAllCounts(const AmArg& args, AmArg& ret) {
       truncate_samples(v, now);
       list<SampleInfo::time_cnt>::iterator v_it = v.begin();
 
-      unsigned int res = 0;
+      long long res = 0;
 
       while (v_it != v.end() && timercmp(&(v_it->time), &to, >))
         v_it++;
@@ -437,7 +437,7 @@ void Monitor::getAllCounts(const AmArg& args, AmArg& ret) {
         }
       }
 
-      ret[s_it->first] = (int)res;
+      ret[s_it->first] = res;
     }
 
   }
